Create pastes routes module
