+++ lib/models/pastes.js
@@ -1,5 +1,5 @@
 const { ObjectID } = require('mongodb');
-const { getCollection } = require('./database');
+const { getCollection } = require('../database');
 
 const collectionName = 'pastes';
